+++ api/fe.js
@@ -1,0 +1,7 @@
+const { generateScript } = require('../forward_engineering/helpers/generateScript');
+const { generateContainerScript } = require('../forward_engineering/helpers/generateContainerScript');
+
+module.exports = {
+	generateScript,
+	generateContainerScript,
+};
